+++ packages/webapp/src/year/Years.tsx
@@ -14,20 +14,36 @@ import { useStoreActions } from '../store/hooks';
 import { useStoreState } from "easy-peasy";
 import { Entry } from '../store/entry-model';
 
+interface YearInfo {
+  year: number,
+  count: number,
+  images: number,
+  videos: number
+}
+
 export const Years = () => {
   const allEntries = useStoreState(state => state.entries.allEntries);
   const history = useHistory();
 
-  const years = useMemo(() => {
+  const yearInfos: YearInfo[] = useMemo(() => {
     const entries: Entry[] = Array.from(allEntries.values());
-    const years = entries.reduce((result, {year}) => {
-      if (result.indexOf(year) < 0) {
-        result.push(year);
+
+    const year2info = entries.reduce((result, {type, date}) => {
+      const year = date.substring(0, 4) || '1970'
+      if (!result[year]) {
+        result[year] = { year: +year, count: 0, images: 0, videos: 0 }
+      }
+      const info = result[year]
+      info.count++
+      switch (type) {
+        case 'image':
+        case 'rawImage': info.images++; break;
+        case 'video': info.videos++
       }
       return result;
     }, [])
-    years.sort((a,b) => b < a ? -1 : 1);
-    return years;
+
+    return Object.values(year2info).sort((a, b) => b.year - a.year)
   }, [allEntries]);
 
   return (
@@ -35,11 +51,15 @@ export const Years = () => {
       <NavBar />
       <h2 style={{marginTop: '40px'}}>Years</h2>
       <ul className="menu">
-        {years.map(year => {
+        {yearInfos.map(({year, count, images, videos}) => {
           return <li className="-list" key={year}>
-            <Link to={`/years/${year}`}>{year}</Link>
-            <a onClick={() => history.push(`/years/${year}?q=image`)}><i className="fas fa-image"></i> <span className="hide-sm">Images</span></a>
-            <a onClick={() => history.push(`/years/${year}?q=video`)}><i className="fas fa-play"></i> <span className="hide-sm">Videos</span></a>
+            <Link to={`/years/${year}`}>{year} - {count} media</Link>
+            { images > 0 &&
+              <a onClick={() => history.push(`/years/${year}?q=image`)}><i className="fas fa-image"></i> <span className="hide-sm">{images} images</span></a>
+            }
+            { videos > 0 &&
+              <a onClick={() => history.push(`/years/${year}?q=video`)}><i className="fas fa-play"></i> <span className="hide-sm">{videos} videos</span></a>
+            }
           </li>
         })}
       </ul>